+++ js/loader.js
@@ -50,9 +50,7 @@ const Loader = (function () {
 	 * @returns {object[]} module data as configured in config
 	 */
 	const getAllModules = function () {
-		console.log(`all modules pre=${JSON.stringify(config.modules)}`);
 		const AllModules = config.modules.filter((module) => (module.module !== undefined) && (MM.getAvailableModulePositions.indexOf(module.position) > -1 || typeof (module.position) === "undefined"));
-		console.log(`all modules=${JSON.stringify(AllModules)}`);
 		return AllModules;
 	};