+++ CHANGELOG.md
@@ -20,7 +20,7 @@ _This release is scheduled to be released on 2024-10-01._
 
 - [weather] Updated `apiVersion` default from 2.5 to 3.0 (#3424)
 - [core] Updated dependencies
-- [core] Allow custom module positions by reading index.html for the defined regions (fixes #3504, fixes runtime and config:check)
+- [core] Allow custom module positions by reading index.html for the defined regions (fixes #3504, fixes runtime and config:check, add testcases)
 
 ### Fixed